FittsState now stores a timestamp.
This allow calculations involving time to be completed using only 2 states, and also allows for the entire test to be logged, analyzed, and played back with accurate timing.

+++ GameManager.pde
@@ -99,7 +99,7 @@ class GameManager {
     newRow.setFloat("activationThreshold", settings.activationThreshold);
     newRow.setFloat("amplitude", consolidatedStats.amplitude);
     newRow.setFloat("width", consolidatedStats.width);
-    newRow.setInt("elapsedTimeMillis", consolidatedStats.elapsedTimeMillis);
+    newRow.setFloat("elapsedTimeMillis", consolidatedStats.elapsedTimeMillis);
     newRow.setFloat("distanceTravelled", consolidatedStats.distanceTravelled);
     newRow.setInt("errors", consolidatedStats.errors);
     newRow.setInt("overShoots", consolidatedStats.overShoots);
